requirements.txt
aliyun-log-cli==0.1.17
aliyun-log-python-sdk==0.6.48.2
antlr4-python3-runtime==4.7.2
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.4.5.1
chardet==3.0.4
click==8.0.1
configobj==5.0.6
dateparser==0.7.4
distlib==0.3.1
Django==3.1.1
django-bootstrap3==14.1.0
docopt==0.6.2
elasticsearch==7.6.0
filelock==3.0.12
Flask==2.0.1
idna==2.9
importlib-metadata==1.7.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.9.5
jsonpath2==0.4.4
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
pip==19.2.3
pipenv==2020.11.15
pipreqs==0.4.10
protobuf==3.11.3
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
setuptools==41.4.0
six==1.14.0
sqlparse==0.3.1
tzlocal==2.0.0
urllib3==1.25.9
virtualenv==20.2.2
virtualenv-clone==0.5.4
Werkzeug==2.0.1
wheel==0.33.6
yarg==0.1.9
zipp==3.2.0
